requirements.txt
# -----------------------------
# Core AI / LLM
# -----------------------------
langchain
langchain-openai
openai
langgraph

# -----------------------------
# API Framework
# -----------------------------
fastapi
uvicorn

# -----------------------------
# Database
# -----------------------------
psycopg2-binary

# -----------------------------
# Data Processing
# -----------------------------
pandas

# -----------------------------
# Environment Variables
# -----------------------------
python-dotenv

# -----------------------------
# Optional (recommended)
# -----------------------------
tqdm